Feat : Download and preprocess KenPom DB data
BODY :

Note that the loop structure where the user may update incompatible
keys may not show every conversion that was made.  Occassionally errors
were made during input and instead of re-applying the changes that had
already been made for the sake of documentation, the cell was simply
re-run and text output covered up.  All teams in the Kaggle competition
data should have a statistics key in the stats_dict.