Feat : Implement NN model and train in a pilot run
BODY :

The initial NN contains 3 fc layers and a classifying layer.  The code
runs from start to until stop.txt file is changed to 'True'.  There
is a bug with NaN in W_1 during the early stages of summarizing, but it
does not affect training.  I think this is a bug with tensorflow / tensor
board.

ref : PILOT_v1